requirements.txt
Django
djangorestframework
psycopg2-binary
django-cors-headers
celery
redis
gunicorn
BeautifulSoup4
lxml
flake8
git+https://github.com/VirtualJudge/VirtualJudgeSpider